requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
openai==1.12.0
httpx==0.27.2
python-dotenv==1.0.1
neo4j==5.18.0
typer==0.9.0
rich==13.7.1
spacy==3.7.4 